Opraveno zpět z Double na Int

+++ Prvocinitele.java
@@ -3,15 +3,15 @@ public class Prvocinitele {
 	public static void main(String[] args) {
 		java.util.Scanner scanner = new java.util.Scanner(System.in);
 		
-		while (scanner.hasNextDouble()) {
-			double cislo = scanner.nextDouble();
+		while (scanner.hasNextInt()) {
+			int cislo = scanner.nextInt();
 			
 			if (cislo <= 1) {
-				System.out.printf("%.2f neresim.\n", cislo);
+				System.out.printf("%d neresim.\n", cislo);
 				continue;
 			}
 			
-			System.out.printf("%.2f =", cislo);
+			System.out.printf("%d =", cislo);
 			int delitel = 2;
 			while (delitel <= cislo) {
 				while ((cislo % delitel) == 0) {
